# docs/swagger.yaml
definitions:
  controller.CommonError:
    properties:
      error:
        type: string
    type: object
  controller.CommonResponse:
    properties:
      info:
        type: string
    type: object
  models.Employee:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.EmployeeRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      name:
        type: string
    type: object
  models.SignUpRequest:
    properties:
      full_name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.TokenPairResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
info:
  contact:
    email: test@test.com
    name: onlineshop api
    url: http://test.com
  title: onlineshop
paths:
  /api/employees:
    get:
      description: getting list of all employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Employee'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: getting employees
      tags:
      - Employees
    post:
      description: creating new employee
      parameters:
      - description: user info
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/models.EmployeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: creating
      tags:
      - Employees
  /api/employees/{id}:
    delete:
      description: delete employee by id
      parameters:
      - description: emp id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: deleting emp
      tags:
      - Employees
    get:
      description: getting an employee by id
      parameters:
      - description: employee id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Employee'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: getting employee
      tags:
      - Employees
    put:
      description: updating user by id
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: user info
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/models.EmployeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      security:
      - BearerAuth: []
      summary: updating
      tags:
      - Employees
  /auth/refresh:
    get:
      description: refresh token pair
      parameters:
      - description: insert refresh token
        in: header
        name: X-Refresh-Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Refresh token pair
      tags:
      - Auth
  /auth/sign-in:
    post:
      description: sign in account
      parameters:
      - description: sign in credentials
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TokenPairResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Sign in
      tags:
      - Auth
  /auth/sign-up:
    post:
      description: create new account
      parameters:
      - description: Information about new account
        in: body
        name: request_body
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.CommonError'
      summary: Registration
      tags:
      - Auth
swagger: "2.0"